Refactor to remove duplication
Create base AndroidSoftware types that are extended by both Library and Applications and reduce duplication, making all feature extensions available to both types.

Lots of other internal cleanup and minor refactoring, remove some helper classes that can be replaced by the base plugin type now.

Marks TODOs around NiA code.
