+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/AndroidSoftwareDependencies.java
@@ -1,0 +1,11 @@
+package org.gradle.api.experimental.android;
+
+import org.gradle.api.artifacts.dsl.Dependencies;
+import org.gradle.api.artifacts.dsl.DependencyCollector;
+
+@SuppressWarnings("UnstableApiUsage")
+public interface AndroidSoftwareDependencies extends Dependencies {
+    DependencyCollector getImplementation();
+    DependencyCollector getRuntimeOnly();
+    DependencyCollector getCompileOnly();
+}
