--- unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/DEFAULT_SDKS.java
@@ -1,7 +1,0 @@
-package org.gradle.api.experimental.android;
-
-public class DEFAULT_SDKS {
-    public static final int JDK = 11;
-    public static final int TARGET_ANDROID_SDK = 34;
-    public static final int MIN_ANDROID_SDK = 21;
-}
