+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/application/AndroidApplicationBuildType.java
@@ -2,15 +2,14 @@ package org.gradle.api.experimental.android.application;
 
 import com.android.build.api.dsl.ApplicationVariantDimension;
 import org.gradle.api.Action;
-import org.gradle.api.experimental.android.extensions.Minify;
-import org.gradle.api.experimental.common.ApplicationDependencies;
+import org.gradle.api.experimental.android.AndroidSoftwareBuildType;
 import org.gradle.api.provider.Property;
 import org.gradle.api.tasks.Nested;
 import org.gradle.declarative.dsl.model.annotations.Configuring;
 import org.gradle.declarative.dsl.model.annotations.Restricted;
 
 @Restricted
-public interface AndroidApplicationBuildType {
+public interface AndroidApplicationBuildType extends AndroidSoftwareBuildType {
     /**
      * @see ApplicationVariantDimension#getApplicationIdSuffix()
      */
@@ -23,22 +22,12 @@ public interface AndroidApplicationBuildType {
     @Restricted
     Property<String> getVersionNameSuffix();
 
-    /**
-     * Dependencies for this build type.
-     */
+    @Override
     @Nested
-    ApplicationDependencies getDependencies();
+    AndroidApplicationDependencies getDependencies();
 
     @Configuring
-    default void dependencies(Action<? super ApplicationDependencies> action) {
+    default void dependencies(Action<? super AndroidApplicationDependencies> action) {
         action.execute(getDependencies());
-    }
-
-    @Nested
-    Minify getMinify();
-
-    @Configuring
-    default void minify(Action<? super Minify> action) {
-        action.execute(getMinify());
     }
 }
