+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/library/AndroidLibraryBuildTypes.java
@@ -1,13 +1,14 @@
 package org.gradle.api.experimental.android.library;
 
 import org.gradle.api.Action;
+import org.gradle.api.experimental.android.AndroidSoftwareBuildTypes;
 import org.gradle.api.tasks.Nested;
 import org.gradle.declarative.dsl.model.annotations.Configuring;
 import org.gradle.declarative.dsl.model.annotations.Restricted;
 
 @Restricted
-public interface AndroidLibraryBuildTypes {
-
+public interface AndroidLibraryBuildTypes extends AndroidSoftwareBuildTypes {
+    @Override
     @Nested
     AndroidLibraryBuildType getDebug();
 
@@ -16,6 +17,7 @@ public interface AndroidLibraryBuildTypes {
         action.execute(getDebug());
     }
 
+    @Override
     @Nested
     AndroidLibraryBuildType getRelease();
