+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/library/AndroidLibraryDependencies.java
@@ -16,9 +16,10 @@
 
 package org.gradle.api.experimental.android.library;
 
+import org.gradle.api.experimental.android.AndroidSoftwareDependencies;
 import org.gradle.api.experimental.common.LibraryDependencies;
 import org.gradle.declarative.dsl.model.annotations.Restricted;
 
 @Restricted
-public interface AndroidLibraryDependencies extends LibraryDependencies {
+public interface AndroidLibraryDependencies extends LibraryDependencies, AndroidSoftwareDependencies {
 }
