+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/nia/package-info.java
@@ -1,2 +1,7 @@
+// TODO: Everything in this package should be moved to the NiA project
+/**
+ * Contains build logic necessary for the Now In Android (NiA) project as
+ * a temporary workaround to support that project using Declarative Gradle.
+ */
 @org.gradle.api.NonNullApi
 package org.gradle.api.experimental.android.nia;
