+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/package-info.java
@@ -1,0 +1,2 @@
+@org.gradle.api.NonNullApi
+package org.gradle.api.experimental.android;
